feature: Extracted some widgets and corrected filter buttons

+++ lib/features/compare/domain/compare_teacher/compare_teacher_optional_qualifications.dart
@@ -6,9 +6,9 @@ part 'compare_teacher_optional_qualifications.g.dart';
 @freezed
 class CompareTeacherOptionalQualifications with _$CompareTeacherOptionalQualifications {
   const factory CompareTeacherOptionalQualifications({
-    @Default(0) int worked,
-    @Default(0) int late,
-    @Default(0) int assistance,
+    @Default(0) double worked,
+    @Default(0) double late,
+    @Default(0) double assistance,
   }) = _CompareTeacherOptionalQualifications;
 
   factory CompareTeacherOptionalQualifications.fromJson(Map<String, dynamic> json) => _$CompareTeacherOptionalQualificationsFromJson(json);
